game: merge parts in libtrx
