+++ src/tr2/global/vars.c
@@ -6,8 +6,6 @@ const char *g_TRXVersion = "TR2X (non-Docker build)";
 
 int32_t g_TRVersion = TR_VERSION; // overriden at runtime when loading a level
 
-int32_t g_OverlayFlag = 1;
-
 int32_t g_PhdPersp = 0;
 int32_t g_PhdTop = 0;
 int32_t g_PhdLeft = 0;
